1. dev on testing framework, and debug workflow for functionality with block testcases. 2. dockerized application 3. defined test pipeline and test cases for block indexing. 4. define callback functions for the various endpoints (although not implemented).

+++ src/BlockchainIndexer.cpp
@@ -170,7 +170,7 @@ bool BlockchainIndexer::getTransactionValue(std::string aTransactionId, int aIdx
             }
         }
 
-        aValue = std::stoi(val);
+        aValue = std::stod(val);
         return true;
     }
     else
